proverbs/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-identity
azure-keyvault
azure-keyvault-secrets
azure-storage-blob
praw
utils